requirements.txt
python-vlc==3.0.12118
yandex_music==1.0.0
gunicorn==20.1.0
paste==3.5.0
deploy==1.9.1
pastedeploy==2.1.1
pystray==0.19.2
shyaml==0.6.2
